requirements.txt
fastapi
uvicorn
openai==0.28
psycopg2-binary
python-dotenv
